refactor skmultiflow utilities and integrate them (finally)
