+++ river/ensemble/boosting.py
@@ -4,7 +4,7 @@ import typing
 
 from river import base, linear_model
 from river.utils.random import poisson
-from river.utils.skmultiflow_utils import normalize_values_in_dict, scale_values_in_dict
+from river.utils.norm import normalize_values_in_dict, scale_values_in_dict
 
 __all__ = ["AdaBoostClassifier"]
