+++ river/tree/nodes/efdtc_nodes.py
@@ -1,7 +1,7 @@
 import copy
 import numbers
 
-from river.utils.skmultiflow_utils import normalize_values_in_dict
+from river.utils.norm import normalize_values_in_dict
 
 from ..splitter.nominal_splitter_classif import NominalSplitterClassif
 from .branch import (
