+++ river/tree/nodes/hatc_nodes.py
@@ -3,7 +3,7 @@ import typing
 
 from river import stats as st
 from river.utils.random import poisson
-from river.utils.skmultiflow_utils import normalize_values_in_dict
+from river.utils.norm import normalize_values_in_dict
 
 from ..utils import do_naive_bayes_prediction
 from .branch import (
